debug

+++ TastierMachine/src/TastierMachine/Machine.hs
@@ -213,7 +213,7 @@ run = do
           let valueList = map (dmem!) [stAddr..endAddr]
           let charList = map (chr . fromIntegral) valueList
           let string = show charList
-          let stringNoQuotes = map (string!!) [1..len]
+          let stringNoQuotes = map (string!!) [1..(fromIntegral len)]
           tell $ [stringNoQuotes]
           put $ machine { rpc = rpc + 1, rtp = rtp - 1 }
           run
